migrate: Bnnaer component
