+++ src/components/links/index.js
@@ -1,4 +1,4 @@
-export { default as HeaderLink } from '@/components/layout';
 export { default as ArrowLink } from '@/components/links/ArrowLink';
 export { default as ButtonLink } from '@/components/links/ButtonLink';
+export { default as HeaderLink } from '@/components/links/HeaderLink';
 export { default as UnstyledLink } from '@/components/links/UnstyledLink';
